🐛 fix: 반복되는 컴포넌트 추출, mock data 추가
