+++ packages/frontend/src/pages/stock-detail/components/MetricItem.tsx
@@ -1,0 +1,13 @@
+export interface MetricItemProps {
+  label: string;
+  value: string;
+}
+
+export const MetricItem = ({ label, value }: MetricItemProps) => {
+  return (
+    <>
+      <span className="display-medium14 text-gray font-bold">{label}</span>
+      <span className="display-medium14 text-dark-gray">{value}</span>
+    </>
+  );
+};
