+++ packages/frontend/src/pages/stock-detail/components/index.ts
@@ -1,1 +1,4 @@
 export * from './TextArea';
+export * from './MetricItem';
+export * from './MetricSection';
+export * from './Title';
