+++ packages/frontend/tsconfig.app.json
@@ -25,5 +25,5 @@
 
     "composite": true
   },
-  "include": ["src"]
+  "include": ["src", "src/**/*.json"]
 }
